Setup a simple application starting point
The project, as it is given to me, could not even be built on my
machine. I could not connect to the dev server. Normally, I'd focus
on fixing that problem, because it's important to follow the existing
development processes. In this example, however, I see that as a red
herring. The time I would spend fixing my dev environment would add
no value to a project that has no future.

Still, this has forced me to commit a couple hours to understanding
webpack configuration. That's valuable. Personally, I'd rather just
use create-react-app and follow those conventions. It would take a
lot of the headache away. I still might do that.