Improve styling of Property component
I usually do styling in iterations, starting with basic layout and
some of the easy stuff. With each iteration, I focus on getting a
particular aspect of the design into a completed state. Basic styles
are better than no styles, and it gives me room to switch to the
api implementation (the fun part) before I grind out the styling
specifics.